test: centralize InputTextBox lookup to handle collapsed mini-window source
Extend UITestHelper.FindInputTextBox to expand the MiniWindow's collapsed
source-text container before the UIA lookup, then swap every direct
FindFirstDescendant("InputTextBox") call site over to the helper. Without
this, TranslationTests.MiniWindow_TranslateWithEnter (and other mini-window
flows in Translation/Phonetic/Dictionary test suites) saw InputTextBox as
offscreen because the container starts showing SourceTextCollapsed after
the sticky-header refactor.
